requirements.txt
fastapi
uvicorn
pymongo
twilio
google-cloud-texttospeech
openai
python-dotenv
